Add equivalent support for catkin tools
Make the command ros.createCatkinPackage, as well as the build and test
tasks work in catkin tools-based workspaces.

The extension will only activate if it detects either of the
.catkin_workspace or .catkin_tools hidden files in the
workspace.rootPath. Users will still have to build their workspace once
before using the extension, and the extension does not provide a way to
change the build system while it is active.

Resolves: #4